Fix 41 audit findings across runtime, validation, and security boundaries.
Hardens multi-input execution, field resolution, expression semantics, compatibility analysis, registry trust, and parser input limits so silent corruption and bypass paths from the adversarial review are closed.

+++ tests/phase_0_9.rs
@@ -296,7 +296,7 @@ fn runtime_builtin_functions_and_rules() {
             vec![
                 RuntimeValue::String("hello".into()),
                 RuntimeValue::Integer(1),
-                RuntimeValue::Integer(4),
+                RuntimeValue::Integer(3),
             ],
             RuntimeValue::String("ell".into()),
         ),
